feat: 결제 확정 로직 임시 구현하여 첨부 refactor: Integer 사용에서 Long 으로 종합 변경 refactor: 공통 API 사용해도 결과값 받을 수 있도록 api-handler.js 수정

+++ src/main/java/com/bootcamp/paymentdemo/domain/order/dto/OrderCreateResponse.java
@@ -10,13 +10,13 @@ public class OrderCreateResponse {
 
     private Long memberId;
     private Long orderId;
-    private Integer totalAmount;
+    private Long totalAmount;
     private String orderNumber;
 
     private OrderCreateResponse(
             Long memberId,
             Long orderId,
-            Integer totalAmount,
+            Long totalAmount,
             String orderNumber
     ) {
         this.memberId = memberId;
@@ -28,7 +28,7 @@ public class OrderCreateResponse {
     public static OrderCreateResponse register(
             Long memberId,
             Long orderId,
-            Integer totalAmount,
+            Long totalAmount,
             String orderNumber
     ) {
         return new OrderCreateResponse(memberId, orderId, totalAmount, orderNumber);
